Splitting out functions for reuse.
