+++ src/crud/data-mapping/get-data.js
@@ -1,0 +1,10 @@
+import _ from 'lodash';
+import getDataFromReq from './get-data-from-req';
+
+export default function getData(rules, req) {
+    if (!rules) {
+        return;
+    }
+    const fromReq = getDataFromReq(rules.fromReq, req);
+    return _.merge({}, rules.static, fromReq);
+}
